requirements.txt
# discord bot
discord
python-dotenv

aiosqlite
aiohttp

openai>=1.0.0